Avoid wildcard imports, extract default exception handler
